wpa_supplicant.conf: Change IEEE 802.1X/EAPOL version
wpa_supplicant is implemented based on IEEE Std 802.1X-2004
which defines EAPOL version 2. However, there are many APs
that do not handle the new version number correctly (they
seem to drop the frames completely). In order to make
wpa_supplicant interoperate with these APs, the version
number is set to 1 by default. This configuration value can
be used to set it to the newversion (2).

Jira-Id: ADM-4540

Signed-off-by: Oleh Cherpak <oleh.cherpak@globallogic.com>
Change-Id: Iec725858fa0a00d6ddab759ffcba98f97b723b78